+++ example/lib/global.dart
@@ -1,5 +1,0 @@
-import 'models/model_a.dart';
-import 'models/model_b.dart';
-
-ModelA modelA = ModelA();
-ModelB modelB = ModelB();