+++ lib/src/annotations/propagate_annotation.dart
@@ -1,0 +1,7 @@
+/// `PropagateAnnotation` provides the `@propagate` annotation for use in your data model classes.
+class PropagateAnnotation
+{
+  const PropagateAnnotation();
+}
+
+const PropagateAnnotation propagate = PropagateAnnotation();